--- aspnetcore/mvc/views/view-components/sample6.x/ViewCompFinal/ViewComponents/PriorityListViewComponent.cs
@@ -1,27 +1,0 @@
-//#define SECOND
-#if SECOND
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.EntityFrameworkCore;
-using ViewComponentSample.Models;
-
-namespace ViewComponentSample.ViewComponents;
-
-public class PriorityListViewComponent : ViewComponent
-{
-    private readonly ToDoContext db;
-
-    public PriorityListViewComponent(ToDoContext context) => db = context;
-
-    public async Task<IViewComponentResult> InvokeAsync(
-    int maxPriority, bool isDone)
-    {
-        var items = await GetItemsAsync(maxPriority, isDone);
-        return View(items);
-    }
-    private Task<List<TodoItem>> GetItemsAsync(int maxPriority, bool isDone)
-    {
-        return db!.ToDo!.Where(x => x.IsDone == isDone &&
-                             x.Priority <= maxPriority).ToListAsync();
-    }
-}
-#endif
