+++ aspnetcore/mvc/views/view-components/sample6.x/ViewCompFinal/ViewComponents/PriorityListViewComponent1.cs
@@ -1,4 +1,4 @@
-#define First
+//#define First
 #if First
 #region snippet1
 using Microsoft.AspNetCore.Mvc;
