+++ aspnetcore/mvc/views/view-components/sample6.x/ViewCompFinal/Views/Shared/Components/PriorityList/PVC.cshtml
@@ -1,0 +1,10 @@
+﻿@model IEnumerable<ViewComponentSample.Models.TodoItem>
+
+<h2> PVC Named Priority Component View</h2>
+<h4>@ViewBag.PriorityMessage</h4>
+<ul>
+    @foreach (var todo in Model)
+    {
+        <li>@todo.Name</li>
+    }
+</ul>
